# spec/openapi.yaml
openapi: 3.1.0
info:
  title: ADP API
  version: 0.3.0
  description: >
    Native and compat-plane HTTP surface for ADP — see docs/pragmatic_mvp.md
    §2.4 for what is in scope and §2.5 for what is deliberately cut.

    Two planes share this file. `/api/v3` is the GitHub-compatible surface, kept
    faithful enough that unmodified `gh` works against it. `/api/adp` is the
    native plane, which has no GitHub analogue: workspaces, candidate sets,
    sessions and checkpoints, the operation log, and evidence bundles.

    Completeness is enforced, not aspirational: server/src/spec-coverage.test.ts
    enumerates the server's real route table and fails if this file and the
    implementation describe different APIs in either direction. Deliberate
    omissions are the git smart-HTTP wire protocol (specified by git itself and
    proxied verbatim), the loopback-only receive-path hook endpoints, and the
    inbound GitHub webhook receiver, whose payload contract is GitHub's.

    `info.version` above is the contract version, and every response carries it
    in an `ADP-API-Version` header — including 401s and 404s, so a generated
    client can pin before it authenticates. The git smart-HTTP routes are the
    exception: they are proxied verbatim and are not part of this contract.
    See docs/api-compatibility.md for what a version bump promises.
servers:
  - url: /
paths:
...
  /api/v3/repos/{owner}:
    post:
      x-required-scope: "repo:write"
      summary: Create a repository
      security:
        - bearerAuth: []
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepoRequest"
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "201":
          description: Created
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Validation failed or repo already exists
  /api/v3/repos/{owner}/{repo}:
    get:
      security:
        - bearerAuth: []
      x-required-scope: "repo:read"
      summary: Get a repository
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "200":
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not found
...
  /api/v3/user/repos:
    post:
      x-required-scope: "repo:write"
      summary: Create a repository owned by the authenticated user (GitHub-shaped).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepoRequest"
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "201":
          description: Created
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Validation failed, or the repository already exists
  /api/v3/orgs/{org}/repos:
    parameters:
      - name: org
        in: path
        required: true
        schema:
          type: string
    post:
      x-required-scope: "repo:write"
      summary: Create a repository owned by an org (GitHub-shaped).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepoRequest"
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "201":
          description: Created
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Validation failed, or the repository already exists
...
  /api/adp/orgs/{orgId}/repos:
    parameters:
      - $ref: "#/components/parameters/orgId"
    get:
      x-required-scope: "repo:read"
      summary: >
        M4-7: every repo in the org with the land requirements that actually
        resolve for it right now (instance ∪ org ∪ repo, the union
        core/repo-policy.ts's resolveLandRequirements computes), each one
        annotated with which layers ask for it. This is the answer to "why does
        this repo require an approval?" that otherwise means reading two YAML
        files out of two repos and knowing how the union works.
      security:
        - bearerAuth: []
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "200":
          description: Repos in the org with their resolved land policy
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrgRepoPolicy"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Token is not scoped to this org, or not a member
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: No such organization
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Validation failed
...
components:
  parameters:
...
    orgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    OrgRepoPolicy:
      type: object
      required: [id, owner, name, default_branch, is_policy_repo, repo_requirements, resolved]
      properties:
        id: { type: string, format: uuid }
        owner: { type: string }
        name: { type: string }
        default_branch: { type: string }
        is_policy_repo: { type: boolean }
        repo_requirements:
          type: array
          items: { type: string, enum: [gates_green, one_approval] }
        resolved:
          type: array
          description: The union in force, each requirement annotated with every layer that asks for it.
          items:
            type: object
            required: [requirement, from]
            properties:
              requirement: { type: string, enum: [gates_green, one_approval] }
              from:
                type: array
                items: { type: string, enum: [instance, org, repo] }
...
    Error:
      type: object
      required: [message]
      description: >
        #97: the one error shape every 4xx carries. `errors` appears on
        validation failures (422) and is a deliberate projection of the
        server's internal validator output — path, message, and a stable
        code — never the validator's raw issue objects, whose shape belongs
        to a library this contract must outlive.
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [path, message, code]
            properties:
              path:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              message:
                type: string
              code:
                type: string
...
    CreateRepoRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        default_branch:
          type: string
          default: main
...